@@ -1,6 +1,24 @@
 source 'http://rubygems.org'
 
-#gem 'rails', '3.0.5'
+gem 'rails', '3.1.2'
+
+gem 'mysql2', '0.4.10'
+
+gem 'rsolr', '1.0.13'
+
+gem 'paperclip', '2.3.8'
+
+gem 'memcache-client', '1.8.5'
+
+gem 'devise', '1.1.8'
+
+gem 'activerecord-mysql2-adapter', '0.0.3'
+
+gem 'sqlite3', '1.3.4'
+
+gem 'jquery-rails', '3.1.5'
+
+gem 'execjs', '2.7.0'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
@@ -17,59 +35,65 @@ source 'http://rubygems.org'
 #gem 'devise'
 
 # BEGIN include specific versions of these gems
-gem 'abstract', '1.0.0'
-gem 'actionmailer', '3.0.5'
-gem 'actionpack', '3.0.5'
-gem 'activemodel', '3.0.5'
-gem 'activerecord', '3.0.5'
-gem 'activerecord-mysql2-adapter', '0.0.3'
-gem 'activeresource', '3.0.5'
-gem 'activesupport', '3.0.5'
-gem 'archive-tar-minitar', '0.5.2'
-gem 'arel', '2.0.10'
-gem 'bcrypt', '3.1.10'
-gem 'bcrypt-ruby', '2.1.4'
-gem 'bigdecimal', '1.1.0'
-gem 'builder', '2.1.2'
-gem 'bundler', '1.11.2'
-gem 'bundler-unload', '1.0.2'
-gem 'climate_control', '0.0.3'
-gem 'cocaine', '0.5.8'
-gem 'devise', '1.1.8'
-gem 'erubis', '2.6.6'
+#gem 'abstract', '1.0.0'
+#gem 'actionmailer', '3.1.2'
+#gem 'actionpack', '3.1.2'
+#gem 'activemodel', '3.1.2'
+#gem 'activerecord', '3.1.2'
+#gem 'activerecord-mysql2-adapter', '0.0.3'
+#gem 'activeresource', '3.1.2'
+#gem 'activesupport', '3.1.2'
+#gem 'archive-tar-minitar', '0.5.2'
+#gem 'arel', '2.0.10'
+#gem 'bcrypt', '3.1.10'
+#gem 'bcrypt-ruby', '2.1.4'
+#gem 'bigdecimal', '1.1.0'
+#gem 'builder', '2.1.2'
+#gem 'bundler', '1.16.6'
+#gem 'bundler-unload', '1.0.2'
+#gem 'climate_control', '0.0.3'
+#gem 'cocaine', '0.5.8'
+#gem 'devise', '1.1.8'
+#gem 'erubis', '2.6.6'
 #gem 'executable-hooks', '1.3.2'
-gem 'gem-wrappers', '1.2.7'
-gem 'hoe', '3.14.2'
-gem 'i18n', '0.7.0'
-gem 'io-console', '0.3'
-gem 'json', '1.5.5'
-gem 'mail', '2.2.20'
-gem 'memcache-client', '1.8.5'
-gem 'mime-types', '1.25.1'
-gem 'mini_portile2', '2.0.0'
-gem 'minitest', '2.5.1'
-gem 'mysql2', '0.4.10'
-gem 'nokogiri', '1.6.7.1'
-gem 'orm_adapter', '0.0.7'
-gem 'paperclip', '2.3.8'
-gem 'passenger', '5.0.23'
-gem 'polyglot', '0.3.5'
+#gem 'gem-wrappers', '1.2.7'
+#gem 'hoe', '3.14.2'
+#gem 'i18n', '0.7.0'
+#gem 'io-console', '0.3'
+#gem 'json', '1.5.5'
+#gem 'mail', '2.2.20'
+#gem 'memcache-client', '1.8.5'
+#gem 'mime-types', '1.25.1'
+#gem 'mini_portile2', '2.0.0'
+#gem 'minitest', '2.5.1'
+#gem 'mysql2', '0.4.10'
+#gem 'nokogiri', '1.6.7.1'
+#gem 'orm_adapter', '0.0.7'
+#gem 'paperclip', '2.3.8'
+#gem 'passenger', '5.0.23'
+#gem 'polyglot', '0.3.5'
 #gem 'rack', '1.6.4'
-gem 'rack-mount', '0.6.14'
-gem 'rack-test', '0.5.7'
-gem 'rails', '3.0.5'
-gem 'railties', '3.0.5'
-gem 'rake', '10.4.2'
-gem 'rcov', '0.9.11'
-gem 'rdoc', '3.9.5'
-gem 'rsolr', '1.0.13'
-gem 'rubygems-bundler', '1.4.4'
-gem 'rvm', '1.11.3.9'
-gem 'thor', '0.14.6'
-gem 'treetop', '1.4.15'
-gem 'tzinfo', '0.3.46'
-gem 'warden', '1.0.3'
+#gem 'rack-mount', '0.6.14'
+#gem 'rack-test', '0.5.7'
+#gem 'rails', '3.1.2'
+#gem 'railties', '3.1.2'
+#gem 'rake', '10.4.2'
+#gem 'rcov', '0.9.11'
+#gem 'rdoc', '3.9.5'
+#gem 'rsolr', '1.0.13'
+#gem 'rubygems-bundler', '1.4.4'
+#gem 'rvm', '1.11.3.9'
+#gem 'thor', '0.14.6'
+#gem 'treetop', '1.4.15'
+#gem 'tzinfo', '0.3.46'
+#gem 'warden', '1.0.3'
 # END include
+
+group :assets do
+  gem 'sass-rails',   "~> 3.1.5"
+  gem 'coffee-rails', "~> 3.1.1"
+  gem 'uglifier',     ">= 1.0.3"
+end
 
 #gem 'rake', '0.8.7'
 #gem 'actionpack', '3.0.5'
