@@ -1,8 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-Litchfield::Application.config.session_store :cookie_store, :key => '_litchfield_session'
+Litchfield::Application.config.session_store :cookie_store, key: '_litchfield_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
-# (create the session table with "rake db:sessions:create")
-# Litchfield.application.config.session_store :active_record_store
+# (create the session table with "rails generate session_migration")
+# Litchfield::Application.config.session_store :active_record_store
