@@ -1,22 +1,42 @@
 Litchfield::Application.configure do
-  # Settings specified here will take precedence over those in config/environment.rb
-
-  # In the development environment your application's code is reloaded on
-  # every request.  This slows down response time but is perfect for development
-  # since you don't have to restart the webserver when you make code changes.
-  config.cache_classes = false
-
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
-
-  # Show full error reports and disable caching
-  config.consider_all_requests_local       = true
-  config.action_view.debug_rjs             = true
-  config.action_controller.perform_caching = true #false
-  config.cache_store = :mem_cache_store, 'localhost', { :namespace => 'litchfield_rails' }
-
-  # Don't care if the mailer can't send
-  config.action_mailer.raise_delivery_errors = false
-  # Print deprecation notices to the Rails logger
-  config.active_support.deprecation = :log
-end
+    # Settings specified here will take precedence over those in config/application.rb
+  
+    # In the development environment your application's code is reloaded on
+    # every request.  This slows down response time but is perfect for development
+    # since you don't have to restart the web server when you make code changes.
+    config.cache_classes = false
+  
+    # Log error messages when you accidentally call methods on nil.
+    config.whiny_nils = true
+  
+    # Show full error reports and disable caching
+    config.consider_all_requests_local       = true
+    config.action_controller.perform_caching = false
+  
+    # Don't care if the mailer can't send
+    config.action_mailer.raise_delivery_errors = false
+  
+    # Print deprecation notices to the Rails logger
+    config.active_support.deprecation = :log
+  
+    # Only use best-standards-support built into browsers
+    config.action_dispatch.best_standards_support = :builtin
+  
+    # Do not compress assets
+    config.assets.compress = false
+  
+    # Expands the lines which load the assets
+    config.assets.debug = true
+  
+    # Defaults to '/assets'
+    config.assets.prefix = '/assets-dev'
+  
+  
+  
+    config.assets.compile = true
+  
+    config.serve_static_assets = false
+  
+    config.assets.digest = false
+  end
+  
